LeadManagement/requirements.txt
# // requirements.txt 

annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
pyodbc==5.2.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
python-dotenv==1.0.0